+++ docs/Doxyfile.in
@@ -5,7 +5,7 @@ PROJECT_BRIEF          = "Hydra plugin for Maya"
 OUTPUT_DIRECTORY       = @CMAKE_CURRENT_BINARY_DIR@/docs/
 INPUT                  = @CMAKE_CURRENT_SOURCE_DIR@/../README.md @CMAKE_CURRENT_SOURCE_DIR@/../src/ @CMAKE_CURRENT_SOURCE_DIR@/../docs
 USE_MDFILE_AS_MAINPAGE = @CMAKE_CURRENT_SOURCE_DIR@/../README.md
-CREATE_SUBDIRS         = YES
+CREATE_SUBDIRS         = NO
 
 SOURCE_BROWSER         = NO
 INLINE_SOURCES         = NO
