+++ src/hdmaya/adapters/aiSkydomeLightAdapter.cpp
@@ -101,6 +101,8 @@ public:
                     .asChar()));
         } else if (paramName == HdLightTokens->enableColorTemperature) {
             return VtValue(false);
+        } else if (paramName == HdLightTokens->textureResource) {
+            return {};
         }
         return {};
     }
